requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dateutil==2.8.2
PyYAML>=6.0
bcrypt>=4.0.0
extra-streamlit-components>=0.1.60
